chore: move type
