+++ src/components/composite/TokenListCard/TokenListCard.types.ts
@@ -1,7 +1,5 @@
 import type { MinimalToken } from 'src/types/tokens';
 
-export type ResponsiveValue<T> = T | { mobile: T; desktop: T };
-
 export enum TokenListCardTokenSize {
   SM = 'sm',
   MD = 'md',
