+++ src/components/composite/TokenListCard/constants.ts
@@ -1,5 +1,5 @@
 import { AvatarSize } from 'src/components/core/AvatarStack/AvatarStack.types';
-import type { ResponsiveValue } from './TokenListCard.types';
+import type { ResponsiveValue } from '@/types/responsive';
 import { TokenListCardTokenSize } from './TokenListCard.types';
 import type { TypographyProps } from '@mui/material/Typography';
