+++ src/types/responsive.ts
@@ -1,0 +1,1 @@
+export type ResponsiveValue<T> = T | { mobile: T; desktop: T };
